+++ health-services/individual/CHANGELOG.md
@@ -1,7 +1,7 @@
 All notable changes to this module will be documented in this file.
 
 ## 1.1.3 
-- Search by uuid added for individual search
+- Added ability to search by user UUID for individual search.
 
 ## 1.1.2
 - upgraded version from beta