+++ health-services/individual/CHANGELOG.md
@@ -1,5 +1,8 @@
 All notable changes to this module will be documented in this file.
 
+## To be released
+- In IndividualSearch, individualId, mobileNumber updated to accept list of search entities instead of single entity
+
 ## 1.1.2
 - upgraded version from beta