+++ health-services/individual/CHANGELOG.md
@@ -1,5 +1,8 @@
 All notable changes to this module will be documented in this file.
 
+## 1.1.4 - 2024-05-10
+- Integrated Boundary v2 functionality
+- 
 ## 1.1.3 
 - Added ability to search by user UUID for individual search.